+++ modules/post-processing/yolo-object-detection/stack.py
@@ -2,12 +2,12 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import logging
-from typing import Any, Optional, cast
+from typing import Any, cast
 
 import aws_cdk.aws_ecr as ecr
 import aws_cdk.aws_iam as iam
 import cdk_nag
-from aws_cdk import Aspects, Duration, RemovalPolicy, Stack, Tags
+from aws_cdk import Aspects, Duration, Stack, Tags
 from cdk_nag import NagPackSuppression, NagSuppressions
 from constructs import Construct, IConstruct
 
@@ -24,7 +24,7 @@ class ObjectDetection(Stack):
         deployment_name: str,
         module_name: str,
         s3_access_policy: str,
-        removal_policy: Optional[RemovalPolicy] = RemovalPolicy.DESTROY,
+        ecr_repository_arn: str,
         stack_description: str,
         **kwargs: Any,
     ) -> None:
@@ -42,15 +42,8 @@ class ObjectDetection(Stack):
 
         dep_mod = f"{project_name}-{deployment_name}-{module_name}"
 
-        self.repository_name = dep_mod
-
-        repo = ecr.Repository(
-            self,
-            id=self.repository_name,
-            repository_name=self.repository_name,
-            removal_policy=removal_policy,
-            auto_delete_images=True if removal_policy == RemovalPolicy.DESTROY else False,
-        )
+        repo = ecr.Repository.from_repository_arn(self, "Repository", repository_arn=ecr_repository_arn)
+        self.repository_name = repo.repository_name
 
         self.image_uri = f"{repo.repository_uri}:latest"
 
@@ -75,7 +68,7 @@ class ObjectDetection(Stack):
 
         self.role = iam.Role(
             self,
-            f"{self.repository_name}-sm-role",
+            f"{repo.repository_name}-sm-role",
             assumed_by=iam.CompositePrincipal(
                 iam.ServicePrincipal("sagemaker.amazonaws.com"),
             ),